:arrow_right: Move form components from configurator folder to their own
Input, NumberInput, Select
