+++ src/components/form/__tests__/NumberInput.spec.js
