+++ src/components/form/form.js
@@ -1,0 +1,11 @@
+import Input from './Input';
+import NumberInput from './NumberInput';
+import Select from './Select';
+
+const form = {
+    Input,
+    NumberInput,
+    Select,
+}
+
+export default form;
