+++ src/components/form/index.js
@@ -1,0 +1,3 @@
+/* eslint-env node */
+const form = require('./form');
+module.exports = form.default || form;
